requirements.txt
netifaces
pigpio>=1.44
RPI.GPIO
luma.oled
adafruit-circuitpython-seesaw
psutil
